requirements.txt
# Processamento de Dados e Geração Sintética
pandas>=2.0.0
numpy>=1.21.0

# Core de Big Data e Machine Learning
pyspark==3.5.0

# Integração com Nuvem (Google Cloud)
google-cloud-storage>=2.14.0

# Documentação Técnica e Diagramas
diagrams>=0.23.0
graphviz>=0.20.0